move Table component to organisms
