+++ apps/client/pages/exams/index.tsx
@@ -2,7 +2,7 @@ import { ErrorMessage } from 'components/atoms/ErrorMessage';
 import { ExamTile } from 'components/atoms/ExamTile';
 import { Loader } from 'components/atoms/Loader';
 import { SearchInput } from 'components/atoms/SearchInput';
-import { Table } from 'components/molecules/Table/Table';
+import { Table } from 'components/organisms/Table/Table';
 import type { NextPage } from 'next';
 import Head from 'next/head';
 import { useState } from 'react';
